Spam mail detection
This repository contains the code and necessary files for a spam detection model built using TensorFlow and deployed with Flask. The app.py file is the Flask application that serves the model, handling requests to predict whether emails are spam. The train_model.py script is used to train the model on the provided dataset, saving the trained model and tokenizer for future use. The saved model (spam_detection_model.h5) and tokenizer (tokenizer.pickle) are essential for the application to process and predict email messages. The requirements.txt file lists all the dependencies needed to run the project, and the spam.csv file contains the labeled dataset used for training the model.